add_documentation_adjust_code: Rename tester adding _user suffix to help users to undertand pattern used with legacy codes.
