+++ kernel/Makefile
@@ -134,7 +134,7 @@ $(NETDATA_APPS): %: %_kern.o
 ${NETDATA_ALL_APPS}: %: %_kern.o
 
 tester: libbpf
-	$(CC) -I../.local_libbpf -I$(LIBBPF)/src -I$(LIBBPF)/include -I$(LIBBPF)/include/uapi -L../.local_libbpf -o legacy_test tester.c -lbpf -lz -lelf
+	$(CC) -I../.local_libbpf -I$(LIBBPF)/src -I$(LIBBPF)/include -I$(LIBBPF)/include/uapi -L../.local_libbpf -o legacy_test tester_user.c -lbpf -lz -lelf
 
 clean:
 	cd $(LIBBPF)/src && make clean
